+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added emultmw calculation to stellarator and fixed power balance errors (Issue #783)
 - Amended fpump* output to match with primary_pumping options.
 - Corrected power crossing the separatrix for stellarators (Issue #787)
+- Changed Connor-Hastie plasma current model to kappa95 and triang95% (Issue #791)
 
 ## Features
 - HTS REBCO model final version implemented